Revert "test now use junit5, add validRowsCols func and tests"
This reverts commit 59fd199ad0e7e1a1a6fc6035576635dbb54b898a.
