+++ .gitignore
@@ -1,4 +1,1 @@
 /bin/
-.idea/
-out/
-*.iml
