+++ src/tests.java
@@ -1,0 +1,22 @@
+
+public class tests {
+
+	public static void main(String[] args) {
+		print_board_tests();
+
+	}
+	
+	public static void print_board_tests(){
+		String expected = "0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0\n0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0\n0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0\n0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0\n0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0\n0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0\n0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0\n0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0\n0 | 0 | 0 | 0 | 0 | 0 | 0 | 0 | 0";
+		int[][] input = {{0,0,0,0,0,0,0,0,0},{0,0,0,0,0,0,0,0,0},{0,0,0,0,0,0,0,0,0},{0,0,0,0,0,0,0,0,0},{0,0,0,0,0,0,0,0,0},{0,0,0,0,0,0,0,0,0},{0,0,0,0,0,0,0,0,0},{0,0,0,0,0,0,0,0,0},{0,0,0,0,0,0,0,0,0}};
+		String output = sudoku_main.convert_board_to_string(input);
+		if(expected.equals(output)){
+			System.out.println("convert_board_to_string worked");
+		}else{
+			System.out.println("convert_board_to_string failed");
+			System.out.println("Expected: \n" + expected);
+			System.out.println("Output: \n" + output);
+		}
+	}
+
+}
